Refactored to use a single, global price source.
All elements now defer to the global mutable.priceSource

mutable. represents the namespace of elements of our system that are
required to change and effect everything. Use of mutable. should be
limited to values which
1) Affect a substantial section of the system.
and
2) Would cause significant complication to implement in a more
functional or declarative manner.

This removes the mkm-source option from all card price elements.