preordain-card-view now has a set preference.
Added the 'set' property to the preordain-card-view and
preordain-card-prices elements. This allows a preference of which set to
display to be passed so it can be selected by preordain-set-prices.

This is a preference, that means it can be ignored if invalid.

The preference is cleared on a preordain-card-prices level after it is
possibly selected.