Added login alongside assistance.
Login is minimalistic and uncomplicated.

A successful login sets the global mutable.user state. In the future it
should redirect to their custom section.

A failed login leverages the helpers to shake the button alongside the
other cues.

An ajax helper was added, dealing with iron-ajax is actually more of a
hassle.

Nothing is wired up to the login yet.