mutable now persists across sessions.
This lets us persist changes across user sessions in a reactive manner.
No polling needed!

Changes are signalled whenever the classes internal state is modified
which causes that state to be serialized and sent to localStorage.

When mutable is created, it attempts to fetch previous state and ensures
that it is valid.

Changes were made to preordain-source-selector and user-login to
accomodate the change.