user-trades now decorates its output with set releases
The point of this is to add context; a player is going to get more info
that the trade happened close to a set release than just a date.

This adds the decorateTrades to the shared functions. Sets are shown
only if they were released between the first and last trades.

Up to three set releases are shown horizontally before the next release
is set onto a newline.