big preordain-trade-view and dependencies update
I went a little overboard...

user-trades now has the isPublic property that is computed in
preordain-trade-view and passed through. It will let use show edit
buttons only for non-public users.

routing now fills in the name parameter for preordain-trade-view if
necessary. This allows the _nameChanged observer to fire in
preordain-trade-view

user-trade-input now progressively shows more of the input as card
details are entered. The submission button is entered only when all
facets are available.

user-trade-input allows submission of comment and time in order to allow
the eventual submission capability to actually append cards to a trade.

user-trade-input is now 20vh taller, it felt a little cramped.