user-trade-input displays status with preordain-status
Adding a trade can take a noticable amount of time due to OPTIONs
requests, the input ensures the user isn't left grasping for context as
to why everything seems to be hanging.

This does create a pleasant spinner cascade when the input is combined
with user-trades; the spinner on the input completes then a spinner on
the trade list goes off. Unintentional but enjoyable.