added pseudo-aspect ratio to mtg-card
The card would stretch to fill the available space previously, this
meant that the art got framed in in the colored border rather than being
the absolute centerpiece.

By enforcing the max-width of the card as the crop size we typically
deal with, an aspect ratio of a typical magic card is enforced as a side
effect. Cleaner than the flexbox hacks we'd otherwise be dealing with.