preordain-set-gallery lazy loads images ala preordian-set-list
Previously a set would load the entirety of its images causing quite the
churn for bandwidth. Now sets load their images in batches dependent on
the size of the display

This copies the method used in preordain-set-list but also simplifies
the gallery viewing process now that we can assume layout is more or
less constant across sets.