preordain-app scrolls to top of content on view change
Previously, when navigating between cards via the typeahead, the scroll
position would be at wherever it was when navigation happened. This is
undesirable as a search for a card means that you want to see the card,
not whatever you eventually stumbled onto for the old card.

Now, navigation, in all cases, causes a scroll to the top of the
content.