preordain-typeahead quality of life improvements
The typeahead was not responsive when results weren't instant or a bad
search was made. Added the pending and notFound templates in response.

Stripped number of arguments passed to typeahead to a limit that is
exposed by the element. Solves some jank when backspacing on mobile.

After a search is made the query should be discarded. That's exactly
what happens now.