applied visibility detection to full quality loads in mtg-card-art
0c3098e719b6030 introduced a more effective method for determining if
element are in view, this leverages that towards mtg-card-art.

mtg-card-art now avoids fetching expensive full card art when it is less
than 10% in the viewport.

This eliminates a crazy amount of jank when performing transitions
between routes due to so many large images being requested.
Additionally, this is a dramatic decrease in page weight for
people just browsing through the top n cards of a set rather than the
entire buffer we preload for them.

Perhaps one of the best improvements for UX in the entire site.