apply ViewNotifyBehavior to preordain-set-list
Previously, the spinner at the bottom of the element was responsible for
checking page position. This required that visibility checking be a
complicated affair.

The behavior allows us to significantly simplify the progressive loading
of set summaries into the gallery. Additionally, we no longer need to
check if this is the active element due to smarter visibility checking.