started tracking scroll position between views
Previously, moving between views meant that scroll position would be
lost. For giant galleries like /sets, this is a nasty loss of context.
This change means we scroll to where the user was previously for each
view.

This has an opt-in behavior of the discard-scroll event and associated
listener of allowing children to notify preordain-app that they have had
a breaking context switch.

Additionally, the polling period was reduced from 50ms to 10ms for
determining when animation completed. The difference is noticeable when
viewed with this solution.

closes #21